Fix: move vision dataclasses to contextualization dataclasses
