+++ cognite/client/_api/vision.py
@@ -1,8 +1,13 @@
 from typing import Any, Dict, List, Optional, Type, Union
 
 from cognite.client._api_client import APIClient
-from cognite.client.data_classes.contextualization import T_ContextualizationJob
-from cognite.client.data_classes.vision import FeatureParameters, InternalId, VisionExtractJob, VisionFeature
+from cognite.client.data_classes.contextualization import (
+    FeatureParameters,
+    InternalId,
+    T_ContextualizationJob,
+    VisionExtractJob,
+    VisionFeature,
+)
 from cognite.client.utils._auxiliary import assert_type, to_camel_case
 from cognite.client.utils._identifier import IdentifierSequence
 
@@ -55,7 +60,7 @@ class VisionAPI(APIClient):
         """Start an asynchronous job to extract features from image files.
 
         Args:
-            features (Union[cognite.client.data_classes.vision.VisionFeature, List[cognite.client.data_classes.vision.VisionFeature]]): The feature(s) to extract from the provided image files.
+            features (Union[VisionFeature, List[VisionFeature]]): The feature(s) to extract from the provided image files.
             file_ids (List[int]): IDs of the image files to analyze. The images must already be uploaded in the same CDF project.
             file_external_ids (List[str]): The external file ids of the image files to analyze.
         Returns:
@@ -65,7 +70,7 @@ class VisionAPI(APIClient):
             Start a job, wait for completion and then get the parsed results::
 
                 >>> from cognite.client import CogniteClient
-                >>> from cognite.client.data_classes.vision import VisionFeature
+                >>> from cognite.client.data_classes.contextualization import VisionFeature
                 >>> c = CogniteClient()
                 >>> extract_job = c.vision.extract(features=VisionFeature.ASSET_TAG_DETECTION, file_ids=[1])
                 >>> extract_job.wait_for_completion()
