+++ tests/tests_integration/test_api/test_vision_extract.py
@@ -2,9 +2,9 @@ import pytest
 
 from cognite.client import CogniteClient
 from cognite.client.data_classes import FileMetadata
-from cognite.client.data_classes.contextualization import JobStatus
-from cognite.client.data_classes.vision import (
+from cognite.client.data_classes.contextualization import (
     FeatureParameters,
+    JobStatus,
     PeopleDetectionParameters,
     VisionExtractJob,
     VisionFeature,
